Convert more longs to shorts

+++ UCR.Core/Models/Plugin.cs
@@ -133,7 +133,7 @@ namespace HidWizards.UCR.Core.Models
             return Profile.GetRuntimeState(stateGuid);
         }
 
-        protected long ReadOutput(int number)
+        protected short ReadOutput(int number)
         {
             return Outputs[number].CurrentValue;
         }
